workers/2_image_processing/vision_api_handler/requirements.txt
google-cloud-vision==3.4.5
boto3==1.34.131
